Updated Linux LED sample
