+++ Source/Meadow.Core.Samples/IO/SerialPort_Echo/app.config.yaml
@@ -1,0 +1,7 @@
+﻿Logging:
+  LogLevel:
+    Default: Trace
+Lifecycle:
+  ResetOnAppFailure: false
+Update:
+  Enabled: false
