+++ Source/Meadow.Core.Samples/OS/Logging/CS/app.config.yaml
@@ -1,0 +1,12 @@
+﻿Logging:
+  LogLevel:
+    Default: Trace
+Lifecycle:
+  ResetOnAppFailure: false
+Update:
+  UpdateServer: mqtt.meadowcloud.co
+  UpdatePort: 8883
+  AuthServer: https://staging.meadowcloud.dev
+  AuthPort: 443
+  Enabled: true
+  UseAuthentication: true
