+++ Source/Meadow.Linux.Samples/Meadow.Linux.Samples.sln
@@ -21,8 +21,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Graphics.MicroGraphics", ".
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "platforms", "platforms", "{BF5BD32F-D249-4329-9943-40B35C92AAC5}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DigitalIOSample", "pi\DigitalIOSample\DigitalIOSample.csproj", "{9D6C707E-DE2D-44A5-B944-CC95EC656E35}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "raspberry pi", "raspberry pi", "{2AF6A39A-8213-4B29-988E-F5C76C6A9B8B}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ads1015_Sample", "pi\Ads1015_Sample\Ads1015_Sample.csproj", "{3908C3EB-7AE2-4847-B2AE-313980897629}"
@@ -41,7 +39,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Meadow.Gtk_Sample", "gtk\Me
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Displays.Gtk", "..\..\..\Meadow.Foundation\Source\Meadow.Foundation.Peripherals\Displays.Gtk\Driver\Displays.Gtk.csproj", "{ABE281CB-A2C7-4754-B296-8978F7BC064B}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "wsl", "wsl", "{7FC14CD3-089F-4BD0-8B2F-A247EC4EDD92}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "wsl2", "wsl2", "{7FC14CD3-089F-4BD0-8B2F-A247EC4EDD92}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sensors.Atmospheric.Ccs811", "..\..\..\Meadow.Foundation\Source\Meadow.Foundation.Peripherals\Sensors.Atmospheric.Ccs811\Driver\Sensors.Atmospheric.Ccs811.csproj", "{FD344949-3C84-4BCF-A62B-46BC1FDE2700}"
 EndProject
@@ -58,6 +56,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MQTTnet", "..\..\..\MQTTnet\Source\MQTTnet\MQTTnet.csproj", "{F60127E8-5BF0-40C7-AEEC-363BB2F16D3E}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Displays.TftSpi", "..\..\..\Meadow.Foundation\Source\Meadow.Foundation.Peripherals\Displays.TftSpi\Driver\Displays.TftSpi.csproj", "{215BDB20-6A6F-46C5-920E-360974BF5B08}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Led_Sample", "pi\Led_Sample\Led_Sample.csproj", "{E527DFB5-243D-47B9-B1B5-17D29A3E7D5D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -105,10 +105,6 @@ Global
 		{02A4D46D-6124-493B-98A3-9FA300FAB7DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{02A4D46D-6124-493B-98A3-9FA300FAB7DF}.Release|Any CPU.Build.0 = Release|Any CPU
 		{02A4D46D-6124-493B-98A3-9FA300FAB7DF}.Release|Any CPU.Deploy.0 = Release|Any CPU
-		{9D6C707E-DE2D-44A5-B944-CC95EC656E35}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9D6C707E-DE2D-44A5-B944-CC95EC656E35}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9D6C707E-DE2D-44A5-B944-CC95EC656E35}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9D6C707E-DE2D-44A5-B944-CC95EC656E35}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3908C3EB-7AE2-4847-B2AE-313980897629}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3908C3EB-7AE2-4847-B2AE-313980897629}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3908C3EB-7AE2-4847-B2AE-313980897629}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -183,6 +179,10 @@ Global
 		{215BDB20-6A6F-46C5-920E-360974BF5B08}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{215BDB20-6A6F-46C5-920E-360974BF5B08}.Release|Any CPU.Build.0 = Release|Any CPU
 		{215BDB20-6A6F-46C5-920E-360974BF5B08}.Release|Any CPU.Deploy.0 = Release|Any CPU
+		{E527DFB5-243D-47B9-B1B5-17D29A3E7D5D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E527DFB5-243D-47B9-B1B5-17D29A3E7D5D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E527DFB5-243D-47B9-B1B5-17D29A3E7D5D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E527DFB5-243D-47B9-B1B5-17D29A3E7D5D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -195,7 +195,6 @@ Global
 		{A8F5D89D-A040-40A6-BEC7-36ADE087D7CD} = {7F70E7D5-BAB0-4BB4-AFE4-7447B74568AA}
 		{1B72626F-B78A-49B1-9099-C26C6A87DE71} = {7F70E7D5-BAB0-4BB4-AFE4-7447B74568AA}
 		{02A4D46D-6124-493B-98A3-9FA300FAB7DF} = {7F70E7D5-BAB0-4BB4-AFE4-7447B74568AA}
-		{9D6C707E-DE2D-44A5-B944-CC95EC656E35} = {2AF6A39A-8213-4B29-988E-F5C76C6A9B8B}
 		{2AF6A39A-8213-4B29-988E-F5C76C6A9B8B} = {BF5BD32F-D249-4329-9943-40B35C92AAC5}
 		{3908C3EB-7AE2-4847-B2AE-313980897629} = {2AF6A39A-8213-4B29-988E-F5C76C6A9B8B}
 		{A4B615DA-9E15-4BF0-AF45-6A9C410BFD1F} = {2AF6A39A-8213-4B29-988E-F5C76C6A9B8B}
@@ -214,6 +213,7 @@ Global
 		{6B73A2F0-743A-4320-A10B-740CE17781AE} = {7F70E7D5-BAB0-4BB4-AFE4-7447B74568AA}
 		{F60127E8-5BF0-40C7-AEEC-363BB2F16D3E} = {7F70E7D5-BAB0-4BB4-AFE4-7447B74568AA}
 		{215BDB20-6A6F-46C5-920E-360974BF5B08} = {7F70E7D5-BAB0-4BB4-AFE4-7447B74568AA}
+		{E527DFB5-243D-47B9-B1B5-17D29A3E7D5D} = {2AF6A39A-8213-4B29-988E-F5C76C6A9B8B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {5CF4A510-8877-4F71-BFB1-15D6AE1BF070}
